Move collections module to its own subpackage
