+++ .github/CHANGELOG.md
@@ -5,11 +5,11 @@
 * Added a circuit drawing feature that provides a text-based representation
   of a QNode instance. It can be invoked via `qnode.draw()`. The user can specify
   to display variable names instead of variable values and choose either an ASCII
-  or Unicode charset. 
+  or Unicode charset.
   [(#446)](https://github.com/XanaduAI/pennylane/pull/446)
 
   Consider the following circuit as an example:
-  ```python
+  ```python3
   @qml.qnode(dev)
   def qfunc(a, w):
       qml.Hadamard(0)
@@ -19,18 +19,20 @@
 
       return qml.expval(qml.PauliZ(0) @ qml.PauliZ(1))
   ```
+
   We can draw the circuit after it has been executed:
-  ```
+
+  ```python
   >>> result = qfunc(2.3, [1.2, 3.2, 0.7])
   >>> print(qfunc.draw())
    0: ──H──╭C────────────────────────────╭C─────────╭┤ ⟨Z ⊗ Z⟩
    1: ─────╰RX(2.3)──Rot(1.2, 3.2, 0.7)──╰RX(-2.3)──╰┤ ⟨Z ⊗ Z⟩
-  >>> print(qfunc.draw(charset="ascii"))  
+  >>> print(qfunc.draw(charset="ascii"))
    0: --H--+C----------------------------+C---------+| <Z @ Z>
    1: -----+RX(2.3)--Rot(1.2, 3.2, 0.7)--+RX(-2.3)--+| <Z @ Z>
-  >>> print(qfunc.draw(show_variable_names=True))  
-   0: ──H──╭C─────────────────────────────╭C─────────╭┤ ⟨Z ⊗ Z⟩ 
-   1: ─────╰RX(a)──Rot(w[0], w[1], w[2])──╰RX(-1*a)──╰┤ ⟨Z ⊗ Z⟩ 
+  >>> print(qfunc.draw(show_variable_names=True))
+   0: ──H──╭C─────────────────────────────╭C─────────╭┤ ⟨Z ⊗ Z⟩
+   1: ─────╰RX(a)──Rot(w[0], w[1], w[2])──╰RX(-1*a)──╰┤ ⟨Z ⊗ Z⟩
   ```
 
 * Added a quantum chemistry package, `pennylane.qchem`, which supports
@@ -92,7 +94,7 @@
   ```
   QNodes that return probabilities fully support autodifferentiation.
 
-* Added the covenience load functions ``qml.from_pyquil``, ``qml.from_quil`` and 
+* Added the covenience load functions ``qml.from_pyquil``, ``qml.from_quil`` and
   ``qml.from_quil_file`` that convert pyquil objects and Quil code to PennyLane
   templates. This feature requires the latest version of the PennyLane-Forest plugin.
   [#459](https://github.com/XanaduAI/pennylane/pull/459)
