+++ pennylane/collections/__init__.py
@@ -1,0 +1,23 @@
+# Copyright 2018-2020 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""
+The collections module. Contains the QNodeCollection and functions
+that process them.
+"""
+from .qnode_collection import QNodeCollection
+
+from .apply import apply
+from .dot import dot
+from .map import map
+from .sum import sum
